dedup trustwallet asset list function
